tests/requirements.txt
pytest
pytest-cov
codecov
flake8-tabs
pytest-flask
blinker
pep8-naming
